RavenDB-18008 refactoring
